+++ lib/snarky/src/camlsnark_c/libsnark-caml/depends/libff/libff/algebra/fields/fp6_2over3.tcc
@@ -219,7 +219,7 @@ Fp6_2over3_model<n, modulus> Fp6_2over3_model<n,modulus>::cyclotomic_exp(const b
     bool found_nonzero = false;
     std::vector<long> NAF = find_wnaf(1, exponent);
 
-    for (long i = NAF.size() - 1; i >= 0; --i)
+    for (long i = static_cast<long>(NAF.size() - 1); i >= 0; --i)
     {
         if (found_nonzero)
         {
